removed redundant demo
