--- sample/app/http_service/handlers.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"net/http"
-)
-
-func conchHandler(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusAccepted)
-}
-
-func oHandler(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
