--- sample/app/httprouter_service/setup_test.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	. "gopkg.in/check.v1"
-	"testing"
-)
-
-func init() {
-	Suite(&filterTestSuite{})
-}
-
-type filterTestSuite struct {
-}
-
-func Test(t *testing.T) { TestingT(t) }
-
-func (s *filterTestSuite) SetUpTest(_ *C) {
-}
