+++ sample/app/synthesis_demo/handlers.go
@@ -5,6 +5,14 @@ import (
 	"net/http"
 )
 
+func conchHandler(w http.ResponseWriter, _ *http.Request) {
+	w.WriteHeader(http.StatusAccepted)
+}
+
+func oHandler(w http.ResponseWriter, _ *http.Request) {
+	w.WriteHeader(http.StatusOK)
+}
+
 func conchRoute(w http.ResponseWriter, _ *http.Request, ps httprouter.Params) {
 	if ps.ByName("ok") != "" {
 		w.WriteHeader(http.StatusAccepted)
